Moved to_tuple to shape_utils
